quick sort + counting sort

+++ src/ua/com/golubov/algo/stepik/test/ArrayInversionServiceTest.java
@@ -3,7 +3,7 @@ package ua.com.golubov.algo.stepik.test;
 import org.junit.Test;
 import ua.com.golubov.algo.stepik.ArrayInversionService;
 
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
@@ -17,7 +17,7 @@ public class ArrayInversionServiceTest {
 
     @Test
     public void testMergeSortAlgo() {
-        AtomicInteger counter = new AtomicInteger(0);
+        AtomicLong counter = new AtomicLong(0);
         int[] sorted = ArrayInversionService.mergeSortWithCounter(new int[]{2, 3, 9, 2, 9}, 0, 4, counter);
         assertArrayEquals(new int[]{2, 2, 3, 9, 9}, sorted);
         assertEquals(2, counter.intValue());
@@ -25,7 +25,7 @@ public class ArrayInversionServiceTest {
 
     @Test
     public void testMergeSortAlgoLargerDataSet() {
-        AtomicInteger counter = new AtomicInteger(0);
+        AtomicLong counter = new AtomicLong(0);
         int[] sorted = ArrayInversionService.mergeSortWithCounter(new int[]{2, 3, 9, 2, 9, 1, 4, 5, 6, 7, -1}, 0, 10, counter);
         assertArrayEquals(new int[]{-1, 1, 2, 2, 3, 4, 5, 6, 7, 9, 9}, sorted);
         assertEquals(25, counter.intValue());
